feat: change history tracker with UI for auditing record changes (#2068)
Add a change history tracking with gem logidze.

Store a history which keeps track of every change made to certain
entities and add a "History" tab to the Reaction/Sample detail pages.
Related entities are displayed in the version history. E.g. Reaction ->
ReactionsSample.
Save timestamp and author of the version with the changes.
A change diff to the previous version can be displayed.
Instead of one version per database, update versions are grouped by the
request id (since some requests can lead to multiple updates for the same record).
These changes are merged to make it more intuitive for the user.

Reverting a record to previous state:
the user can update selected fields of a record to a former values.
Note that this is not a roll-back and the applied reverting changes are tracked
as well.

Tracked entities:
- Sample
  - Residue
  - ElementalComposition
- Reaction
- ReactionsSample
- ResearchPlan
  - ResearchPlansWellplate
  - ResearchPlanMetadata
- Screen
- Wellplate
  - Well
- Container
- Attachment

(replace #1253)

-----

- implement Logidze
  - catch whodunit and wrap it into each DB update/insert transaction
- api for fetching/reverting history
- fetchers, reverters, and serializers in app/service/versioning

- add json column to store the history for each tracked model
- build specific `gem-fx` versioned sql logging functions
- added triggers on update insert for tracked elements
 a `gem-fx` versioned function is used to generate the default trigger for each model
 instead of using `gem-fx` versioned individual triggers since all triggers are identical
- NB migration squashed and timestamped forward
- ⚠️ migration initiate history object for all records of the tracked models
- migration rollback confirmed:
NB: back-rolling the logging function also roll-back all triggers since they depends on it
- added soft deletion on container
- some update on older migration

- history accessible in an additional tab in each element-detail card
- specific svg formatter and json
- history tab syncs with properties - reloads on click
- change-table using AgGridReact

---------

Co-authored-by: Martin Schneider <m.schneider@cortona.de>
Co-authored-by: VadimKeller <v.keller@cortona.de>
Co-authored-by: Martin Schneider <mschneider85@users.noreply.github.com>
Co-authored-by: VadimKeller <VadimKeller@users.noreply.github.com>
Co-authored-by: Adrian Herrmann <adrian.herrmann@kit.edu>